Update Vendor, Buyer view for User

+++ src/main/java/vn/gpay/gsmart/core/api/pcontract/PContract_getbysearch_request.java
@@ -5,8 +5,8 @@ import vn.gpay.gsmart.core.base.RequestBase;
 public class PContract_getbysearch_request extends RequestBase {
 	public String productbuyer_code;
 	public String po_code;
-	public Integer orgbuyerid_link;
-	public Integer orgvendorid_link;
+	public Long orgbuyerid_link;
+	public Long orgvendorid_link;
 	public String contractbuyer_code;
 	public Integer contractbuyer_yearfrom;
 	public Integer contractbuyer_yearto;
